Rename Tts component and fix broken references
