--- pages/tts-test.js
@@ -1,5 +1,0 @@
-import Tts from "../components/tts";
-
-export default function TtsTest() {
-  return <Tts>Welcome to Alis Learning</Tts>;
-}
